+++ R/add.R
@@ -1,10 +1,9 @@
 #' Add molecular formula assignments
 #' @rdname addAssignments
-#' @description Add molecular formaul assignments to an `Analysis` class object.
-#' @param analysis S4 object of class `Analysis`
-#' @param assignment S4 object of class `Assignment`
-#' @return An S4 object of class `Analysis`
-#' @details Molecular formula assignments are added to the `Analysis` pre-treated data 
+#' @description Add molecular formula assignments to an `Analysis` S4 class object.
+#' @param analysis an object of S4 class `Analysis`
+#' @param assignment an object of S4 class `Assignment` that includes assignemnts for the *m/z* features specified for argument `analysis`
+#' @return An object of S4 class `Analysis` that includes the molecular formulas in the feature names for the pre-treated data.
 #' @export
 
 setGeneric('addAssignments',
@@ -12,7 +11,7 @@ setGeneric('addAssignments',
            standardGeneric('addAssignments'))
 
 #' @rdname addAssignments
-#' @importFrom MFassign assignedData
+#' @importFrom assignments assignedData
 
 setMethod('addAssignments',signature = c('Analysis','Assignment'),
           function(analysis,assignment){